Moved all classes to the proper file structure folders and created an IStemmer interface
