+++ src/main/java/dk/sdu/se_f22/BrandModule/Stemming/IStemmer.java
@@ -1,0 +1,5 @@
+package dk.sdu.se_f22.BrandModule.Stemming;
+
+public interface IStemmer {
+  
+}
