+++ src/main/java/dk/sdu/se_f22/BrandModule/Stemming/Word.java
